requirements.txt
python-telegram-bot==20.*
pydantic>=2.6
filelock>=3.12
jinja2>=3.1
openai>=1.14
google-generativeai>=0.8
apscheduler>=3.10
pytest>=8.0
